add S2VertexLike.
